Renamed Swap to SwapERC20 and SwapAny to Swap, including utils, librairies. Updated SwapERC20, Swap, Maker and Wrapper tests accordingly
