+++ source/swap/contracts/interfaces/IKittyCoreTokenTransfer.sol
