+++ tools/libraries/src/SwapERC20.ts
@@ -1,9 +1,9 @@
 import { ethers, BigNumber, ContractTransaction } from 'ethers'
 import { chainIds, chainNames } from '@airswap/constants'
-import { Order } from '@airswap/typescript'
+import { OrderERC20 } from '@airswap/typescript'
 import { SwapERC20 as SwapContract } from '@airswap/swap-erc20/typechain/contracts'
 import { SwapERC20__factory } from '@airswap/swap-erc20/typechain/factories/contracts'
-import { orderToParams } from '@airswap/utils'
+import { orderERC20ToParams } from '@airswap/utils'
 
 import * as swapDeploys from '@airswap/swap-erc20/deploys.js'
 
@@ -33,7 +33,7 @@ export class SwapERC20 {
   }
 
   public async check(
-    order: Order,
+    order: OrderERC20,
     senderWallet: string,
     signer?: ethers.providers.JsonRpcSigner
   ): Promise<Array<string>> {
@@ -47,13 +47,13 @@ export class SwapERC20 {
     }
     const [count, errors] = await contract.check(
       senderWallet,
-      ...orderToParams(order)
+      ...orderERC20ToParams(order)
     )
     return this.convertToArray(count, errors)
   }
 
   public async swap(
-    order: Order,
+    order: OrderERC20,
     sender?: ethers.providers.JsonRpcSigner
   ): Promise<ContractTransaction> {
     let contract = this.contract
@@ -64,11 +64,14 @@ export class SwapERC20 {
         contract = contract.connect(sender)
       }
     }
-    return await contract.swap(sender.getAddress(), ...orderToParams(order))
+    return await contract.swap(
+      sender.getAddress(),
+      ...orderERC20ToParams(order)
+    )
   }
 
   public async swapAnySender(
-    order: Order,
+    order: OrderERC20,
     sender?: ethers.providers.JsonRpcSigner
   ): Promise<ContractTransaction> {
     let contract = this.contract
@@ -81,12 +84,12 @@ export class SwapERC20 {
     }
     return await contract.swapAnySender(
       sender.getAddress(),
-      ...orderToParams(order)
+      ...orderERC20ToParams(order)
     )
   }
 
   public async swapLight(
-    order: Order,
+    order: OrderERC20,
     sender?: ethers.providers.JsonRpcSigner
   ): Promise<ContractTransaction> {
     let contract = this.contract
@@ -97,7 +100,7 @@ export class SwapERC20 {
         contract = contract.connect(sender)
       }
     }
-    return await contract.swapLight(...orderToParams(order))
+    return await contract.swapLight(...orderERC20ToParams(order))
   }
 
   private convertToArray(count: BigNumber, errors: Array<string>) {
